+++ pkg/vvm/impl_cfg.go
@@ -12,9 +12,9 @@ import (
 
 	"github.com/voedger/voedger/pkg/iprocbus"
 	"github.com/voedger/voedger/pkg/iprocbusmem"
+	"github.com/voedger/voedger/pkg/isecretsimpl"
 	"github.com/voedger/voedger/pkg/istorage"
 	"github.com/voedger/voedger/pkg/istructs"
-	"github.com/voedger/voedger/pkg/itokensjwt"
 	commandprocessor "github.com/voedger/voedger/pkg/processors/command"
 	"github.com/voedger/voedger/pkg/router"
 )
@@ -53,15 +53,9 @@ func NewVVMDefaultConfig() VVMConfig {
 			logger.Info("using istoragemem")
 			return istorage.ProvideMem(), nil
 		},
+		SecretsReader: isecretsimpl.ProvideSecretReader(),
 	}
 	return res
-}
-
-func (tsr *testISecretReader) ReadSecret(name string) ([]byte, error) {
-	if name == SecretKeyJWTName {
-		return itokensjwt.SecretKeyExample, nil
-	}
-	return tsr.realSecretReader.ReadSecret(name)
 }
 
 func (cfg *VVMConfig) addProcessorChannel(cg iprocbusmem.ChannelGroup, t ProcessorChannelType) {
