+++ pkg/vvm/types.go
@@ -127,6 +127,7 @@ type VVMConfig struct {
 	// test and FederationURL contains port -> the port will be relaced with the actual VVMPort
 	FederationURL       *url.URL
 	ActualizerStateOpts []state.ActualizerStateOptFunc
+	SecretsReader       isecrets.ISecretReader
 }
 
 type resultSenderErrorFirst struct {
@@ -140,8 +141,4 @@ type VoedgerVM struct {
 	*VVM
 	vvmCtxCancel func()
 	vvmCleanup   func()
-}
-
-type testISecretReader struct {
-	realSecretReader isecrets.ISecretReader
 }
